separate tags and color palette into 2 pages/directories
