+++ tools/sidekick/palette/palette.css
@@ -1,0 +1,75 @@
+main {
+    font-family: system-ui, Arial;
+    padding: 20px;
+    font-size: 1.3rem;
+    line-height: 1.2em;
+}
+
+main .intro {
+    font-size: 1.2rem;
+}
+
+#palette {
+    margin: 0 0 1rem;
+}
+
+#palette p {
+    font-size: 13px;
+    line-height: 1.5;
+}
+
+#palette ul {
+    list-style-type: none;
+    display: grid;
+    grid-template-columns: repeat(4, 1fr);
+    grid-template-rows: masonry;
+    margin: 0;
+    padding: 0;
+    background: aliceblue;
+    font-size: 14px;
+}
+
+#palette ul li {
+    position: relative;
+    padding: .5rem;
+}
+
+#palette ul li:hover {
+    background-color: white;
+    cursor: pointer;
+}
+
+#palette ul li.copied {
+    background-color: #cccc;
+}
+
+#palette ul li.copied::after {
+    position: absolute;
+    content: attr(data-name) ' copied';
+    top: 8px;
+    right: 6px;
+    display: block;
+    padding: 4px;
+    background-color: rgb(234 255 234);
+    font-size: 12px;
+    border: 1px solid rgb(95 176 95);
+    line-height: 12px;
+}
+
+#palette ul .label p {
+    margin: 0;
+    padding: 0;
+}
+
+#palette ul .label p.value {
+    margin: 0;
+}
+
+#palette ul .swatch {
+    margin: 0;
+    padding: 0;
+    height: 25px;
+    width: 100%;
+    display: block;
+    border: 1px solid black;
+}
