+++ tools/sidekick/palette/palette.html
@@ -1,0 +1,20 @@
+<!doctype html>
+<html>
+  <head>
+    <title>Sling - Color Palette</title>
+    <meta name="viewport" content="width=device-width, initial-scale=1" />
+    <script src="/tools/sidekick/colors/palette.js" type="module"></script>
+    <link rel="stylesheet" href="/tools/sidekick/colors/palette.css" />
+  </head>
+  <body>
+    <main>
+      <div class="intro">
+        <h1>Color Variables Palette</h1>
+        <p>Click to copy color variable for text, backgrounds, and buttons</p>
+      </div>
+      <section id="palette">
+        <ul></ul>
+      </section>
+    </main>
+  </body>
+</html>
