+++ tools/sidekick/tagger/tagger.css
@@ -1,4 +1,3 @@
-
 main {
     font-family: system-ui, Arial;
     padding: 20px;
@@ -156,69 +155,4 @@ span.psep {
 .offscreen {
     position: absolute;
     top: -1000px;
-}
-
-#palette {
-    margin: 0 0 1rem;
-
-    p {
-        font-size: 13px;
-        line-height: 1.5;
-    }
-
-    ul {
-        list-style-type: none;
-        display: grid;
-        grid-template-columns: repeat(4, 1fr);
-        grid-template-rows: masonry; /* stylelint-disable-line */
-        margin: 0;
-        padding: 0;
-        background: aliceblue;
-        font-size: 14px;
-
-        li {
-            position: relative;
-            padding: .5rem;
-        }
-
-        li:hover {
-            background-color: white;
-            cursor: pointer;
-        }
-
-        li.copied {
-            background-color: #cccc;
-
-            &::after {
-                position: absolute;
-                content: attr(data-name) ' copied';
-                top: 8px;
-                right: 6px;
-                display: block;
-                padding: 4px;
-                background-color: rgb(234 255 234);
-                font-size: 12px;
-                border: 1px solid rgb(95 176 95);
-                line-height: 12px;
-            }
-        }
-
-        .label p {
-            margin: 0;
-            padding: 0;
-        }
-
-        .label p.value {
-            margin: 0;
-        }
-
-        .swatch {
-            margin: 0;
-            padding: 0;
-            height: 25px;
-            width: 100%;
-            display: block;
-            border: 1px solid black;
-        }
-    }
 }
