+++ tools/sidekick/tagger/tagger.html
@@ -8,6 +8,8 @@
   <head>
     <title>Sling - Tag Picker</title>
     <meta name="viewport" content="width=device-width, initial-scale=1" />
+    <!-- Import DA App SDK 
+    <script src="https://da.live/nx/utils/sdk.js" type="module"></script> -->
     <script src="/tools/sidekick/tagger/tagger.js" type="module"></script>
     <link rel="stylesheet" href="/tools/sidekick/tagger/tagger.css" />
   </head>
@@ -16,12 +18,10 @@
     <div id="tags-data"></div>
     <main>
       <div class="intro">
-        <h1 id="pick-your-tags">Copy tags and color variables</h1>
-        <p>Use this page to pick the tags for use in page metadata, and copy the color variables for use in color schemes.</p>
-        <h3>Tags for page metadata</h3>
+        <h1 id="pick-your-tags">Pick your tags</h1>
         <p>
           Once you click on a tag, it will populate at the bottom of the page.
-          When you have picked all the tags, click “copy” and then paste the
+          When you have picked all the tags, click "copy" and then paste the
           information directly in your page
         </p>
       </div>
@@ -36,12 +36,6 @@
           <button class="clear">Clear</button>
         </div>
       </div>
-      <hr>
-      <section id="palette">
-        <h3>Color Variables Palette</h3>
-        <p>Click to copy color variable for text, backgrounds, and buttons</p>
-        <ul></ul>
-      </section>
     </main>
     <textarea class="offscreen" id="copybuffer"></textarea>
   </body>
