+++ tools/sidekick/tagger/tagger.js
@@ -1,5 +1,4 @@
-import { getTags, getPalette } from '../../../aemedge/scripts/tags.js';
-import { createTag } from '../../../aemedge/scripts/utils.js';
+import { getTags } from '../../../aemedge/scripts/tags.js';
 
 let selectedOrder = [];
 
@@ -43,7 +42,6 @@ function initTaxonomy(tags) {
       html += '</div>';
     });
   }
-  // results.innerHTML = html;
   results.insertAdjacentHTML('afterbegin', html);
 }
 
@@ -70,7 +68,7 @@ function toggleTag(target) {
   const category = target.closest('.category')
     ?.querySelector('h2')?.textContent;
   const subcategory = target.closest('.subcategory')
-    ?.querySelector('h2')?.textContent; // Assuming category title is in h2
+    ?.querySelector('h2')?.textContent;
 
   const tagIdentifier = {
     title,
@@ -79,7 +77,7 @@ function toggleTag(target) {
   };
 
   if (target.classList.contains('selected')) {
-    selectedOrder.push(tagIdentifier); // Add to the selection order
+    selectedOrder.push(tagIdentifier);
   } else {
     selectedOrder = selectedOrder.filter(
       (item) => item.title !== title || item.category !== category,
@@ -99,7 +97,6 @@ function displaySelected() {
     category,
     subcategory,
   }) => {
-    // Find the category element
     let categories;
     if (subcategory) categories = document.querySelectorAll('#results .subcategory');
     else categories = document.querySelectorAll('#results .category');
@@ -138,63 +135,20 @@ function displaySelected() {
   copybuffer.value = toCopyBuffer.join(', ');
 }
 
-/* color palette rendering */
-function clickToCopyList(items) {
-  items.forEach((item) => {
-    item.addEventListener('click', () => {
-      // Get the attribute you want to copy
-      const attribute = 'data-name';
-      const value = item.getAttribute(attribute);
-      // Copy the attribute value to the clipboard
-      navigator.clipboard.writeText(value)
-        .then(() => {
-          item.classList.add('copied');
-          setTimeout(() => {
-            item.classList.remove('copied');
-          }, 2000);
-        })
-        .catch((err) => {
-          // eslint-disable-next-line no-console
-          console.error('Failed to copy attribute:', err);
-        });
-    });
-  });
-}
-
-async function initPalette() {
-  const palette = await getPalette();
-  if (!palette) return;
-  const palletList = document.querySelector('#palette > ul');
-  palette.forEach((color) => {
-    const brandName = color['brand-name'];
-    const colorValue = color['color-value'];
-    const uses = color['application']; // eslint-disable-line
-    const swatch = createTag('div', { class: 'swatch', style: `background: ${colorValue};` });
-    const label = createTag('div', { class: 'label' }, `<p><strong>${brandName}</strong></p><p>Uses: ${uses}</p><p class="value">${colorValue}</p>`);
-    const colorElem = createTag('li', { class: brandName, 'data-color': colorValue, 'data-name': brandName }, label);
-    colorElem.prepend(swatch);
-    palletList.append(colorElem);
-  });
-  const items = palletList.querySelectorAll('li');
-  if (items) clickToCopyList(items);
-}
-
 async function init() {
   const tags = await getTags();
   initTaxonomy(tags);
-  initPalette();
+
   const selEl = document.getElementById('selected');
   const copyButton = selEl.querySelector('button.copy');
   copyButton.addEventListener('click', () => {
     const copyText = document.getElementById('copybuffer');
     navigator.clipboard.writeText(copyText.value);
-
     copyButton.disabled = true;
   });
 
   const clearButton = selEl.querySelector('button.clear');
   clearButton.addEventListener('click', () => {
-    // Remove the 'filtered' class from all tags
     document.querySelectorAll('#results .tag')
       .forEach((tag) => {
         tag.closest('.path')
@@ -202,7 +156,6 @@ async function init() {
           .remove('filtered');
       });
 
-    // Remove the 'selected' class from all selected tags
     document.querySelectorAll('.selected')
       .forEach((selectedTag) => {
         selectedTag.classList.remove('selected');
@@ -228,4 +181,5 @@ async function init() {
     }
   });
 }
+
 init();
